+++ src/test/kotlin/apply/application/EvaluationServiceTest.kt
@@ -1,105 +1,71 @@
 package apply.application
 
-import apply.EVALUATION_TITLE1
-import apply.EVALUATION_TITLE2
-import apply.EVALUATION_TITLE3
 import apply.createEvaluation
 import apply.createRecruitment
-import apply.domain.evaluation.Evaluation
 import apply.domain.evaluation.EvaluationRepository
+import apply.domain.evaluation.getById
 import apply.domain.evaluationItem.EvaluationItemRepository
-import apply.domain.recruitment.Recruitment
 import apply.domain.recruitment.RecruitmentRepository
+import io.kotest.core.spec.style.BehaviorSpec
+import io.kotest.matchers.booleans.shouldBeFalse
+import io.kotest.matchers.shouldBe
 import io.mockk.Runs
+import io.mockk.clearAllMocks
 import io.mockk.every
-import io.mockk.impl.annotations.MockK
 import io.mockk.just
-import org.assertj.core.api.Assertions.assertThat
-import org.junit.jupiter.api.BeforeEach
-import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.assertAll
-import support.test.UnitTest
-import java.util.Optional
+import io.mockk.mockk
+import support.test.spec.afterRootTest
 
-@UnitTest
-internal class EvaluationServiceTest {
-    @MockK
-    private lateinit var recruitmentRepository: RecruitmentRepository
+class EvaluationServiceTest : BehaviorSpec({
+    val recruitmentRepository = mockk<RecruitmentRepository>()
+    val evaluationRepository = mockk<EvaluationRepository>()
+    val evaluationItemRepository = mockk<EvaluationItemRepository>()
 
-    @MockK
-    private lateinit var evaluationRepository: EvaluationRepository
+    val evaluationService = EvaluationService(evaluationRepository, evaluationItemRepository, recruitmentRepository)
 
-    @MockK
-    private lateinit var evaluationItemRepository: EvaluationItemRepository
+    Given("다수의 모집과 모집에 대한 평가가 있는 경우") {
+        val recruitment1 = createRecruitment(title = "모집1", id = 1L)
+        val recruitment2 = createRecruitment(title = "모집2", id = 2L)
+        val evaluation1 = createEvaluation(title = "평가1", recruitmentId = 1L, beforeEvaluationId = 0L, id = 1L)
+        val evaluation2 = createEvaluation(title = "평가2", recruitmentId = 1L, beforeEvaluationId = 1L, id = 2L)
+        val evaluation3 = createEvaluation(title = "평가3", recruitmentId = 2L, beforeEvaluationId = 0L, id = 3L)
 
-    private lateinit var evaluationService: EvaluationService
-    private lateinit var recruitments: List<Recruitment>
-    private lateinit var preCourseEvaluation: Evaluation
-    private lateinit var firstEvaluation: Evaluation
-    private lateinit var secondEvaluation: Evaluation
-    private lateinit var evaluations: List<Evaluation>
+        every { evaluationRepository.findAll() } returns listOf(evaluation1, evaluation2, evaluation3)
+        every { recruitmentRepository.getOne(1L) } returns recruitment1
+        every { recruitmentRepository.getOne(2L) } returns recruitment2
+        every { evaluationRepository.getById(1L) } returns evaluation1
+        every { evaluationRepository.getById(2L) } returns evaluation2
 
-    @BeforeEach
-    internal fun setUp() {
-        evaluationService = EvaluationService(evaluationRepository, evaluationItemRepository, recruitmentRepository)
+        When("모든 평가를 모집과 함께 조회하면") {
+            val actual = evaluationService.findAllWithRecruitment()
 
-        recruitments = listOf(
-            createRecruitment(recruitable = false),
-            createRecruitment(recruitable = true)
-        )
-
-        preCourseEvaluation = createEvaluation(title = EVALUATION_TITLE1, beforeEvaluationId = 0L, id = 1L)
-
-        firstEvaluation = createEvaluation(title = EVALUATION_TITLE2, beforeEvaluationId = 1L, id = 2L)
-
-        secondEvaluation = createEvaluation(title = EVALUATION_TITLE3, beforeEvaluationId = 2L, id = 3L)
-
-        evaluations = listOf(preCourseEvaluation, firstEvaluation, secondEvaluation)
+            Then("모집명과 평가명을 확인할 수 있다") {
+                actual[0] shouldBe EvaluationResponse(evaluation1, "모집1", 1L, "이전 평가 없음", 0L)
+                actual[1] shouldBe EvaluationResponse(evaluation2, "모집1", 1L, "평가1", 1L)
+                actual[2] shouldBe EvaluationResponse(evaluation3, "모집2", 2L, "이전 평가 없음", 0L)
+            }
+        }
     }
 
-    @Test
-    fun `평가와 모집 정보를 함께 제공한다`() {
-        every { evaluationRepository.findAll() } returns evaluations
-        every { recruitmentRepository.getOne(any()) } returns recruitments[0]
-        every { evaluationRepository.findById(1L) } returns Optional.of(evaluations[0])
-        every { evaluationRepository.findById(2L) } returns Optional.of(evaluations[1])
+    Given("특정 평가가 이전 평가인 다른 평가가 있는 경우") {
+        val evaluation1 = createEvaluation(beforeEvaluationId = 10L, id = 1L)
+        val evaluation2 = createEvaluation(beforeEvaluationId = 1L, id = 2L)
+        val evaluation3 = createEvaluation(beforeEvaluationId = 1L, id = 3L)
 
-        val findAllWithRecruitment = evaluationService.findAllWithRecruitment()
-
-        assertAll(
-            { assertThat(2L).isEqualTo(findAllWithRecruitment[1].id) },
-            { assertThat(firstEvaluation.title).isEqualTo(findAllWithRecruitment[1].title) },
-            { assertThat(firstEvaluation.description).isEqualTo(findAllWithRecruitment[1].description) },
-            { assertThat(recruitments[0].title).isEqualTo(findAllWithRecruitment[1].recruitmentTitle) },
-            { assertThat(firstEvaluation.recruitmentId).isEqualTo(findAllWithRecruitment[1].recruitmentId) },
-            { assertThat(preCourseEvaluation.title).isEqualTo(findAllWithRecruitment[1].beforeEvaluationTitle) },
-            { assertThat(preCourseEvaluation.id).isEqualTo(findAllWithRecruitment[1].beforeEvaluationId) }
-        )
-    }
-
-    @Test
-    fun `삭제된 평가를 이전 평가로 가지는 평가의 이전 평가를 초기화한다`() {
-        every { evaluationRepository.findAll() } returns evaluations
         every { evaluationRepository.deleteById(any()) } just Runs
+        every { evaluationRepository.findAll() } returns listOf(evaluation1, evaluation2, evaluation3)
 
-        evaluationService.deleteById(2L)
+        When("특정 평가를 삭제하면") {
+            evaluationService.deleteById(evaluation1.id)
 
-        assertThat(0L).isEqualTo(evaluations[2].beforeEvaluationId)
+            Then("이전 평가가 없도록 다른 평가가 재설정된다") {
+                evaluation2.hasBeforeEvaluation().shouldBeFalse()
+                evaluation3.hasBeforeEvaluation().shouldBeFalse()
+            }
+        }
     }
 
-    @Test
-    fun `삭제된 평가를 이전 평가로 가지는 평가들의 이전 평가를 초기화한다`() {
-        val thirdEvaluation = createEvaluation(beforeEvaluationId = 2L, id = 4L)
-
-        val evaluationsWithDuplicatedBeforeEvaluationId: List<Evaluation> =
-            listOf(*evaluations.toTypedArray(), thirdEvaluation)
-
-        every { evaluationRepository.findAll() } returns evaluationsWithDuplicatedBeforeEvaluationId
-        every { evaluationRepository.deleteById(any()) } just Runs
-
-        evaluationService.deleteById(2L)
-
-        assertThat(0L).isEqualTo(evaluationsWithDuplicatedBeforeEvaluationId[2].beforeEvaluationId)
-        assertThat(0L).isEqualTo(evaluationsWithDuplicatedBeforeEvaluationId[3].beforeEvaluationId)
+    afterRootTest {
+        clearAllMocks()
     }
-}
+})
