+++ src/test/kotlin/apply/application/MailTargetServiceTest.kt
@@ -1,6 +1,5 @@
 package apply.application
 
-import apply.EVALUATION_ID
 import apply.createEvaluationAnswer
 import apply.createEvaluationTarget
 import apply.createUser
@@ -12,36 +11,30 @@ import apply.domain.evaluationtarget.EvaluationStatus.WAITING
 import apply.domain.evaluationtarget.EvaluationTargetRepository
 import apply.domain.user.UserRepository
 import apply.domain.user.findAllByEmailIn
+import io.kotest.core.spec.style.BehaviorSpec
+import io.kotest.matchers.collections.shouldBeEmpty
+import io.kotest.matchers.collections.shouldHaveSize
+import io.kotest.matchers.shouldBe
+import io.mockk.clearAllMocks
 import io.mockk.every
-import io.mockk.impl.annotations.MockK
+import io.mockk.mockk
 import io.mockk.verify
-import org.assertj.core.api.Assertions.assertThat
-import org.junit.jupiter.api.BeforeEach
-import org.junit.jupiter.api.Test
-import support.test.UnitTest
+import support.test.spec.afterRootTest
 
-@UnitTest
-class MailTargetServiceTest {
-    @MockK
-    private lateinit var evaluationTargetRepository: EvaluationTargetRepository
+class MailTargetServiceTest : BehaviorSpec({
+    val evaluationTargetRepository = mockk<EvaluationTargetRepository>()
+    val userRepository = mockk<UserRepository>()
 
-    @MockK
-    private lateinit var userRepository: UserRepository
+    val mailTargetService = MailTargetService(evaluationTargetRepository, userRepository)
 
-    private lateinit var mailTargetService: MailTargetService
+    Given("특정 평가에 대한 평가 대상자가 있는 경우") {
+        val evaluationId = 1L
 
-    @BeforeEach
-    internal fun setUp() {
-        mailTargetService = MailTargetService(evaluationTargetRepository, userRepository)
-    }
-
-    @Test
-    fun `평가 상태에 따라 (null, 전체) 메일 발송 대상들의 이메일 정보를 불러온다`() {
         every { evaluationTargetRepository.findAllByEvaluationId(any()) } returns listOf(
-            createEvaluationTarget(userId = 1L, evaluationStatus = WAITING),
-            createEvaluationTarget(userId = 2L, evaluationStatus = PENDING),
-            createEvaluationTarget(userId = 3L, evaluationStatus = PASS),
-            createEvaluationTarget(userId = 4L, evaluationStatus = FAIL)
+            createEvaluationTarget(evaluationId = evaluationId, userId = 1L, evaluationStatus = WAITING),
+            createEvaluationTarget(evaluationId = evaluationId, userId = 2L, evaluationStatus = PENDING),
+            createEvaluationTarget(evaluationId = evaluationId, userId = 3L, evaluationStatus = PASS),
+            createEvaluationTarget(evaluationId = evaluationId, userId = 4L, evaluationStatus = FAIL)
         )
         every { userRepository.findAllById(any()) } returns listOf(
             createUser(id = 1L, email = "waiting@email.com"),
@@ -49,82 +42,117 @@ class MailTargetServiceTest {
             createUser(id = 3L, email = "pass@email.com"),
             createUser(id = 4L, email = "fail@email.com")
         )
-        val actual = mailTargetService.findMailTargets(EVALUATION_ID)
-        assertThat(actual).hasSize(4)
+
+        When("해당 평가의 모든 평가 대상자에 대한 이메일 정보를 조회하면") {
+            val actual = mailTargetService.findMailTargets(evaluationId)
+
+            Then("모든 평가 대상자의 이메일 정보를 확인할 수 있다") {
+                actual shouldHaveSize 4
+            }
+        }
     }
 
-    @Test
-    fun `평가 상태에 따라 (PASS) 메일 발송 대상들의 이메일 정보를 불러온다`() {
+    Given("특정 평가에 합격한 평가 대상자가 있는 경우") {
+        val evaluationId = 1L
+        val user = createUser(id = 3L, email = "pass@email.com")
+
         every { evaluationTargetRepository.findAllByEvaluationIdAndEvaluationStatus(any(), any()) } returns listOf(
-            createEvaluationTarget(userId = 3L, evaluationStatus = PASS)
+            createEvaluationTarget(evaluationId = evaluationId, userId = user.id, evaluationStatus = PASS)
         )
-        every { userRepository.findAllById(any()) } returns listOf(
-            createUser(id = 3L, email = "pass@email.com")
-        )
-        val actual = mailTargetService.findMailTargets(EVALUATION_ID, PASS)
-        assertThat(actual).hasSize(1)
-        assertThat(actual[0].email).isEqualTo("pass@email.com")
+        every { userRepository.findAllById(any()) } returns listOf(user)
+
+        When("해당 평가에 합격한 모든 평가 대상자의 이메일 정보를 조회하면") {
+            val actual = mailTargetService.findMailTargets(evaluationId, PASS)
+
+            Then("평가 대상자의 이름 및 이메일을 확인할 수 있다") {
+                actual shouldHaveSize 1
+                actual[0] shouldBe MailTargetResponse(user.email, user.name)
+            }
+        }
     }
 
-    @Test
-    fun `평가 상태에 따라 (FAIL) 메일 발송 대상들의 이메일 정보를 불러온다`() {
+    Given("특정 평가에 탈락한 평가 대상자가 있는 경우") {
+        val evaluationId = 1L
+        val user = createUser(id = 2L, email = "fail@email.com")
+
         every { evaluationTargetRepository.findAllByEvaluationIdAndEvaluationStatus(any(), any()) } returns listOf(
-            createEvaluationTarget(userId = 2L, evaluationStatus = FAIL)
+            createEvaluationTarget(evaluationId = evaluationId, userId = user.id, evaluationStatus = FAIL)
         )
-        every { userRepository.findAllById(any()) } returns listOf(
-            createUser(id = 2L, email = "fail@email.com")
-        )
-        val actual = mailTargetService.findMailTargets(EVALUATION_ID, FAIL)
-        assertThat(actual).hasSize(1)
-        assertThat(actual[0].email).isEqualTo("fail@email.com")
+        every { userRepository.findAllById(any()) } returns listOf(user)
+
+        When("해당 평가에 탈락한 모든 평가 대상자의 이메일 정보를 조회하면") {
+            val actual = mailTargetService.findMailTargets(evaluationId, FAIL)
+
+            Then("평가 대상자의 이름 및 이메일을 확인할 수 있다") {
+                actual shouldHaveSize 1
+                actual[0] shouldBe MailTargetResponse(user.email, user.name)
+            }
+        }
     }
 
-    @Test
-    fun `평가 상태에 따라 (FAIL) 메일 발송 시 과제 미제출자는 대상이 되지 않는다`() {
+    Given("특정 평가에 보류 중인 평가 대상자가 있는 경우") {
+        val evaluationId = 1L
+        val user = createUser(id = 2L, email = "waiting@email.com")
+
+        every { evaluationTargetRepository.findAllByEvaluationIdAndEvaluationStatus(any(), any()) } returns listOf(
+            createEvaluationTarget(evaluationId = evaluationId, userId = user.id, evaluationStatus = WAITING)
+        )
+        every { userRepository.findAllById(any()) } returns listOf(user)
+
+        When("해당 평가에 보류 중인 모든 평가 대상자의 이메일 정보를 조회하면") {
+            val actual = mailTargetService.findMailTargets(evaluationId, WAITING)
+
+            Then("평가 대상자의 이름 및 이메일을 확인할 수 있다") {
+                actual shouldHaveSize 1
+                actual[0] shouldBe MailTargetResponse(user.email, user.name)
+            }
+        }
+    }
+
+    Given("특정 평가에 평가를 받지 않고 탈락한 평가 대상자가 있는 경우") {
+        val evaluationId = 1L
+
         every { evaluationTargetRepository.findAllByEvaluationIdAndEvaluationStatus(any(), any()) } returns listOf(
             createEvaluationTarget(
+                evaluationId = evaluationId,
                 userId = 1L,
                 evaluationStatus = FAIL,
-                evaluationAnswers = EvaluationAnswers(emptyList())
+                evaluationAnswers = EvaluationAnswers(emptyList()),
             ),
             createEvaluationTarget(
+                evaluationId = evaluationId,
                 userId = 2L,
                 evaluationStatus = FAIL,
                 evaluationAnswers = EvaluationAnswers(listOf(createEvaluationAnswer(score = 0)))
             )
         )
         every { userRepository.findAllById(any()) } returns emptyList()
-        mailTargetService.findMailTargets(EVALUATION_ID, FAIL)
-        verify { userRepository.findAllById(emptyList()) }
+
+        When("해당 평가에 탈락한 모든 평가 대상자의 이메일 정보를 조회하면") {
+            val actual = mailTargetService.findMailTargets(evaluationId, FAIL)
+
+            Then("해당 평가 대상자의 이름 및 이메일을 확인할 수 없다") {
+                verify { userRepository.findAllById(emptyList()) }
+                actual.shouldBeEmpty()
+            }
+        }
     }
 
-    @Test
-    fun `평가 상태에 따라 (WAITING) 메일 발송 대상들의 이메일 정보를 불러온다`() {
-        every { evaluationTargetRepository.findAllByEvaluationIdAndEvaluationStatus(any(), any()) } returns listOf(
-            createEvaluationTarget(userId = 1L, evaluationStatus = WAITING)
-        )
-        every { userRepository.findAllById(any()) } returns listOf(
-            createUser(id = 1L, email = "waiting@email.com")
-        )
-        val actual = mailTargetService.findMailTargets(EVALUATION_ID, WAITING)
-        assertThat(actual).hasSize(1)
-        assertThat(actual[0].email).isEqualTo("waiting@email.com")
+    Given("특정 이메일을 가진 회원이 없는 경우") {
+        val email = "test1@email.com"
+
+        every { userRepository.findAllByEmailIn(any()) } returns emptyList()
+
+        When("해당 이메일로 이메일 정보를 조회하면") {
+            val actual = mailTargetService.findAllByEmails(listOf(email))
+
+            Then("이름이 비어있는 것을 확인할 수 있다") {
+                actual[0] shouldBe MailTargetResponse(email, null)
+            }
+        }
     }
 
-    @Test
-    fun `이메일 중에서 회원에 해당하는 이메일이 있으면 (회원이름, 이메일)을, 회원에 해당하는 이메일이 없으면 (공백, 이메일)을 반환한다`() {
-        val users = listOf(
-            createUser(name = "회원1", email = "test1@email.com"),
-            createUser(name = "회원3", email = "test3@email.com")
-        )
-        val emails = listOf("test1@email.com", "test2@email.com", "test3@email.com")
-        val mailTargetResponses = listOf(
-            MailTargetResponse("test1@email.com", "회원1"),
-            MailTargetResponse("test3@email.com", "회원3"),
-            MailTargetResponse("test2@email.com")
-        )
-        every { userRepository.findAllByEmailIn(emails) } returns users
-        val actual = mailTargetService.findAllByEmails(emails)
-        assertThat(actual).isEqualTo(mailTargetResponses)
+    afterRootTest {
+        clearAllMocks()
     }
-}
+})
